+++ src/context.ts
@@ -1,5 +1,4 @@
-
-import * as _ from "lodash";
+import { PluginContext } from "rollup";
 
 export interface IContext
 {
@@ -17,6 +16,11 @@ export enum VerbosityLevel
 	Debug,
 }
 
+function getText (message: string | (() => string)): string {
+	return typeof message === "string" ? message : message();
+}
+
+/** mainly to be used in options hook, but can be used in other hooks too */
 export class ConsoleContext implements IContext
 {
 	constructor(private verbosity: VerbosityLevel, private prefix: string = "")
@@ -27,27 +31,67 @@ export class ConsoleContext implements IContext
 	{
 		if (this.verbosity < VerbosityLevel.Warning)
 			return;
-		console.log(`${this.prefix}${_.isFunction(message) ? message() : message}`);
+		console.log(`${this.prefix}${getText(message)}`);
 	}
 
 	public error(message: string | (() => string)): void
 	{
 		if (this.verbosity < VerbosityLevel.Error)
 			return;
-		console.log(`${this.prefix}${_.isFunction(message) ? message() : message}`);
+		console.log(`${this.prefix}${getText(message)}`);
 	}
 
 	public info(message: string | (() => string)): void
 	{
 		if (this.verbosity < VerbosityLevel.Info)
 			return;
-		console.log(`${this.prefix}${_.isFunction(message) ? message() : message}`);
+		console.log(`${this.prefix}${getText(message)}`);
 	}
 
 	public debug(message: string | (() => string)): void
 	{
 		if (this.verbosity < VerbosityLevel.Debug)
 			return;
-		console.log(`${this.prefix}${_.isFunction(message) ? message() : message}`);
+		console.log(`${this.prefix}${getText(message)}`);
+	}
+}
+
+/** cannot be used in options hook (which does not have this.warn and this.error), but can be in other hooks */
+export class RollupContext implements IContext
+{
+	constructor(private verbosity: VerbosityLevel, private bail: boolean, private context: PluginContext, private prefix: string = "")
+	{
+	}
+
+	public warn(message: string | (() => string)): void
+	{
+		if (this.verbosity < VerbosityLevel.Warning)
+			return;
+		this.context.warn(`${getText(message)}`);
+	}
+
+	public error(message: string | (() => string)): void
+	{
+		if (this.verbosity < VerbosityLevel.Error)
+			return;
+
+		if (this.bail)
+			this.context.error(`${getText(message)}`);
+		else
+			this.context.warn(`${getText(message)}`);
+	}
+
+	public info(message: string | (() => string)): void
+	{
+		if (this.verbosity < VerbosityLevel.Info)
+			return;
+		console.log(`${this.prefix}${getText(message)}`);
+	}
+
+	public debug(message: string | (() => string)): void
+	{
+		if (this.verbosity < VerbosityLevel.Debug)
+			return;
+		console.log(`${this.prefix}${getText(message)}`);
 	}
 }
